Clean up real_api provider backend tests + move them into tests/
